Correct datatype for maxLen (See [ 2172888 ])

+++ httpAdapter.c
@@ -779,7 +779,7 @@ static int doHttpRequest(CommHndl conn_fd)
          cp = &hdr[15];
          cp += strspn(cp, " \t");
          inBuf.content_length = atol(cp);
-         int maxLen;
+         long maxLen;
          getControlNum("httpMaxContentLength", &maxLen);
          if((maxLen) && (inBuf.content_length > maxLen)) {
             genError(conn_fd, &inBuf, 413, "Request Entity Too Large", NULL);
